+++ lib/dataset/include/scipp/dataset/map_view.h
@@ -44,6 +44,11 @@ auto slice_map(const Sizes &sizes, const T &map, const Slice &params) {
       if (value.dims()[params.dim()] == sizes[params.dim()]) {
         out[key] = value.slice(params);
       } else { // bin edge
+        if (params.stride() != 1)
+          throw except::SliceError(
+              "Object has bin-edges along dimension " +
+              to_string(params.dim()) + " so slicing with stride " +
+              std::to_string(params.stride()) + " != 1 is not valid.");
         const auto end = params.end() == -1 ? params.begin() + 2
                                             : params.begin() == params.end()
                                                   ? params.end()
@@ -146,8 +151,8 @@ public:
 
   Dict slice(const Slice &params) const;
   std::tuple<Dict, Dict> slice_coords(const Slice &params) const;
-  void validateSlice(const Slice s, const Dict &dict) const;
-  [[maybe_unused]] Dict &setSlice(const Slice s, const Dict &dict);
+  void validateSlice(const Slice &s, const Dict &dict) const;
+  [[maybe_unused]] Dict &setSlice(const Slice &s, const Dict &dict);
 
   void rename(const Dim from, const Dim to);