+++ lib/python/bind_slice_methods.h
@@ -38,12 +38,10 @@ auto from_py_slice(const T &source,
   const auto size = dim_extent(source, dim);
   if (!indices.compute(size, &start, &stop, &step, &slicelength))
     throw py::error_already_set();
-  if (step != 1)
-    throw std::runtime_error("Step must be 1");
   if (slicelength == 0) {
     stop = start; // Propagate vanishing slice length downstream.
   }
-  return Slice(dim, start, stop);
+  return Slice(dim, start, stop, step);
 }
 
 template <class View> struct SetData {